+++ src/main/kotlin/com/manimdsl/runtime/utlityFunctions.kt
@@ -1,4 +1,4 @@
-package com.manimdsl
+package com.manimdsl.runtime
 private val WRAP_LINE_LENGTH = 50
 
 fun wrapCode(code: MutableList<String>): MutableList<MutableList<String>> {
